Changed keys used to identify reporters in example
The keys in the listed example could possibly be mistaken for reserved words.  Changed the keys to make it more obvious that the keys were defined by the user and to prevent possible confusion around creating multiple file reporters.  Added a link to the reporter interface section of the API as a quick reference.